update smoke test scripts
