+++ .github/workflows/nightly.yml
@@ -60,11 +60,31 @@ jobs:
           body: >
             This PR is auto-generated by a nightly GitHub action. 
             It should automatically be merged if tests pass.
-      
-      - name: Mark Pull Request for Auto-Merge
-        if: steps.createpr.outputs.pull-request-operation == 'created'
-        uses: peter-evans/enable-pull-request-automerge@v1
+
+  smoke-sync:
+    if: github.repository_owner == 'withastro'
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out code using Git
+        uses: actions/checkout@v2
+      - name: Set Node version to 16
+        uses: actions/setup-node@v2
         with:
-          token: ${{ secrets.NIGHTLY_PERSONAL_GITHUB_TOKEN }}
-          pull-request-number: ${{ steps.createpr.outputs.pull-request-number }}
-          merge-method: squash
+          node-version: 16
+          cache: 'yarn'
+      - name: Install dependencies
+        run: yarn install --frozen-lockfile --ignore-engines --ignore-scripts
+      - name: Sync smoke tests
+        run: node scripts/smoke/sync.js
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      - name: Create Pull Request
+        id: createpr
+        uses: peter-evans/create-pull-request@v3
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+          commit-message: '[ci] update smoke tests (remote)'
+          title: '[ci] update smoke tests (remote)'
+          body: >
+            This PR is auto-generated by a nightly GitHub action. 
+            It should automatically be merged if tests pass.
